story #15813 refactor(iam): client data transfer objets
