+++ cas/cas-server/src/test/java/fr/gouv/vitamui/cas/authentication/LoginPwdAuthenticationHandlerTest.java
@@ -1,14 +1,14 @@
 package fr.gouv.vitamui.cas.authentication;
 
 import fr.gouv.vitamui.cas.util.Constants;
+import fr.gouv.vitamui.commons.api.domain.UserDto;
 import fr.gouv.vitamui.commons.api.enums.UserStatusEnum;
 import fr.gouv.vitamui.commons.api.enums.UserTypeEnum;
 import fr.gouv.vitamui.commons.api.exception.BadRequestException;
 import fr.gouv.vitamui.commons.api.exception.InvalidAuthenticationException;
 import fr.gouv.vitamui.commons.api.exception.TooManyRequestsException;
+import fr.gouv.vitamui.iam.common.dto.cas.LoginRequestDto;
 import fr.gouv.vitamui.iam.openapiclient.CasApi;
-import fr.gouv.vitamui.iam.openapiclient.domain.LoginRequestDto;
-import fr.gouv.vitamui.iam.openapiclient.domain.UserDto;
 import jakarta.servlet.http.HttpServletRequest;
 import org.apereo.cas.authentication.Credential;
 import org.apereo.cas.authentication.PreventedException;
